JS: Rework Client/Vendor division
Clean the messy spaghetti code I did while doing stuff for Clients and Vendors. We trade off some code complexity for less code, which can be easily understood if you read the header of utilities.js
Beware, this method saves the user info inside a variable: it's not usable for modifying the userlist; we fallback to the old method in case.

Signed-off-by: Gianmarco Rengucci <rengucci.gianmarco@gmail.com>